+++ Userland/Libraries/LibUnicode/Locale.h
@@ -85,46 +85,6 @@ enum class Style : u8 {
     Numeric,
 };
 
-struct NumberGroupings {
-    u8 primary_grouping_size { 0 };
-    u8 secondary_grouping_size { 0 };
-};
-
-enum class StandardNumberFormatType : u8 {
-    Decimal,
-    Currency,
-    Accounting,
-    Percent,
-    Scientific,
-};
-
-enum class CompactNumberFormatType : u8 {
-    DecimalLong,
-    DecimalShort,
-    CurrencyUnit,
-    CurrencyShort,
-};
-
-struct NumberFormat {
-    enum class Plurality : u8 {
-        Other,
-        Zero,
-        Single,
-        One,
-        Two,
-        Few,
-        Many,
-    };
-
-    u8 magnitude { 0 };
-    u8 exponent { 0 };
-    Plurality plurality { Plurality::Other };
-    StringView zero_format {};
-    StringView positive_format {};
-    StringView negative_format {};
-    Vector<StringView> identifiers {};
-};
-
 struct ListPatterns {
     StringView start;
     StringView middle;
@@ -188,10 +148,6 @@ Optional<StringView> get_locale_script_mapping(StringView locale, StringView scr
 Optional<StringView> get_locale_currency_mapping(StringView locale, StringView currency, Style style);
 Vector<StringView> get_locale_key_mapping(StringView locale, StringView keyword);
 Optional<StringView> get_number_system_symbol(StringView locale, StringView system, StringView symbol);
-Optional<NumberGroupings> get_number_system_groupings(StringView locale, StringView system);
-Optional<NumberFormat> get_standard_number_system_format(StringView locale, StringView system, StandardNumberFormatType type);
-Vector<NumberFormat> get_compact_number_system_formats(StringView locale, StringView system, CompactNumberFormatType type);
-Vector<NumberFormat> get_unit_formats(StringView locale, StringView unit, Style style);
 Optional<ListPatterns> get_locale_list_patterns(StringView locale, StringView type, StringView style);
 
 Optional<StringView> resolve_language_alias(StringView language);
@@ -203,8 +159,5 @@ Optional<StringView> resolve_subdivision_alias(StringView subdivision);
 Optional<LanguageID> add_likely_subtags(LanguageID const& language_id);
 Optional<LanguageID> remove_likely_subtags(LanguageID const& language_id);
 String resolve_most_likely_territory(LanguageID const& language_id, StringView territory_alias);
-
-Optional<NumberFormat> select_pattern_with_plurality(Vector<NumberFormat> const& formats, double number);
-Optional<String> augment_currency_format_pattern(StringView currency_display, StringView base_pattern);
 
 }
